requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
python-dotenv==1.0.0
email-validator==2.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
bcrypt==4.1.2
PyJWT==2.8.0
Flask-Mail==0.9.1
Flask-Admin==1.6.1
Flask-Security-Too==5.3.2
Flask-Limiter==3.5.0
